[cm] Pull out the `Model` dependency from routing
Having `Model` in any of the routing code is viral. Often times, it's
easier to have the routing code return the source `Realm` of a
capability rather than its `AbsoluteMoniker`. No model look up is
needed, since the `Realm` is already found.

The only strange bit is that in order to bind to a `Realm`, the
`Realm`'s parent chain is walked until the root is reached. At this
point, each Realm can be bound to, descending back towards the target
`Realm`.

Change-Id: I99851730b0ef8d520b313427c5bcd4655047b0fd