[kazoo] Avoid causing unnecessary rebuilds when output files are unchanged
My motivation for this change is to reduce incremental build times
when working on experimental syscalls, especially build times when
switching between a local branch and tip-of-tree.

Without this CL, any change to zircon/vdso/*.fidl triggers a rebuild
of most of the codebase, because the timestamp of zircon/syscalls.h
will get updated.  With this CL, I can mark new syscalls as
"[testonly]", and only files that #include zircon/testonly-syscalls.h
will get rebuilt.

That is achieved by only updating timestamps when output files'
contents are changed.  Ninja will notice when rerunning kazoo does not
modify the output files' timestamps.

Drop error printing from kazoo's copy of ReadFileToString() because
WriteFileIfChanged() shouldn't print anything when the file doesn't
exist.  The only existing non-test call site of ReadFileToString()
doesn't need it to print an error because the call site prints an
error itself.

Bug: 45802
Test: new test case in kazoo-test
 * unit tests run locally with:
     fx set bringup.x64
     fx ninja -C out/default.zircon/ host-x64-linux-clang/obj/tools/kazoo/kazoo-test
     out/default.zircon/host-x64-linux-clang/obj/tools/kazoo/kazoo-test
 * tested manually with:
     touch zircon/vdso/syscall.fidl
     fx build
     + observe much fewer targets being rebuilt
Change-Id: Ic72ad3245e3f05bd70327bfea716842d47a3b400