[components] expect_events runs in new task
This CL changes EventSource::expect_events to validate events in
a new task. The result of the task is returned via a oneshot channel
back to the main task.

This allows expect_events to be non-blocking relative to main test code.

This CL also addresses a potential race in input_session_tests that runs
the input session concurrently with expect_events. In principle, it is possible for
event subscription to happen after starting the session, missing some
events. This CL fixes the problem by ensuring event subscription happens
first and waiting for expectations happens after the sessions starts.

Change-Id: I27f0275976a04242e44a6884240bcf9efe02eeba