[system-tests] OTA should verify a reboot actually occurred
When we switched over to explicitly calling the system_updater, we forgot
to use the "write to a temp file to see if a reboot actually occurred"
technique we used in `Device.TriggerSystemOta()`. This patch creates a
new function, `Device.ExpectReboot()` that writes to a temp file, runs
a closure that should cause the device to reboot, and verifies the temp
file no longer exists.

Test: This is a test, and therefore doesn't need tests written to test
the logic. I manually confirmed the OTA tests still work.

Change-Id: Idbcff2a43c890ad68a224482e8dd60ea04862d03