[components] EventSink => EventLog
As part of clarifying the Events API, this CL renames EventSink to
EventLog, and "sink" analogies to be more explicit. In particular,
soak_events => record_events.

Additionally, a TODO is addressed in the EventLog implementation:
once the EventLog goes out of scope, recording stops.

Public facing documentation is also updated with this CL.

Change-Id: Ia6649e803f931975775bea9f48dd7547880b701e