[fbl] Move inline_array to ulib
fbl::InlineArray<T, size> provides a fixed-sized array container,
where data is stored inline (in the instance of InlineArray) when there
are less than |size| elements and out-of-line when there are more.

Move it from the kernel-only part of FBL to ulib, where both user and
kernel code can use it.

Change-Id: Ideb4f938860d4d253c92c28aaeeb94157e11d8ca