[debugger] Add more exception info.
Adds the platform exception information to the exception notification.

This is currently unused. It will be used in a future pass to display
additional exception information.

Convert the stop notification parameters in zxdb to a StopInfo struct
with the new information on it. This will allow us to add more exception
info in the future without such a large change.

Change-Id: I16e361b832d958e30a9f2d46f3c41a8efcadc7e9