[components][migration] Migrate fonts to components v2
This change moves fonts.cmx from the `sys` realm to a child of `core`.
The tests have already been migrated to v2.

We keep fonts.cmx in the package because it is included in some external
integration tests. Once those external tests are migrated to v2, we can
remove it.

Tested: Verified that text still appears

Change-Id: Ia5fe1dbce68335eb91204066f0e5b45c8702861c