[fidl] Assert correct response thread in debug
WeakStubController (and related) aren't thread-safe.  In debug only,
assert that usages of WeakStubController are from the correct thread
(creation thread).

This indirectly also asserts that relevant usages of PendingResponse are
from the correct thread, intentionally without preventing a curry of
PendingResponse to another thread and back before WeakStubController is
actually used.

At least for now, it seems like ThreadChecker from FXL would be more
than we need here.

In release, we have the field, initialized to thrd_t{} and left that
way.  In debug, we don't complain if thread_ == thrd_t{}, just in case.

Tested: CQ, fx make-integration-patch
Change-Id: I71f71728d1221e85a749fe69dbd49fc9626da3ca