[watchdog][cmdline] Provide a cmdline option to early-disable a WDT
Provide a new kernel command line option which allows a user to tell
the kernel to disarm and disable any watchdog passed to it by the
bootloader via the ZBI, if at all possible.  For systems whose
bootloader is arming the watchdog timer, this can provide the ability
disable the watchdog if needed in order to perform early kernel
debugging which might require not petting the dog for extended periods
of time before it can be disabled using the kernel console.

Change-Id: Idd04aeffcc646b826df99416f9fdd07ee10d4bc8