[identity][modular] Unhook Fuchsia Identity from modular, step 1.
This is the first in a sequence of steps discussed over email to remove
Identity from modular now it no longer has active clients.

Step 1 discards the TokenManager given to AgentContextImpl and returns
failures for all the TokenManager methods AgentContextImpl implements.
We will let this soak a while and verify it doesn't break any clients
we were unaware of.

Bug: 45946
Change-Id: Ice466a8dfa005db88833a8b642de4e2d9e2f0082