[kernel] Fix alignment issues in "array" special sections
The special sections that are used as arrays must have precise
bounds symbols and be densely packed with no more alignment padding
than an actual array would have.

The compiler is normally free to overalign any variable as an
optimization.  Importantly, the GNU `aligned` attribute (via the
`__ALIGNED` macro) only requests *at least* that might alignment
and does not prevent the compiler from choosing to do more.

Instead, use the `alignas` keyword on each array-section object to
tell the compiler to align it to exactly its type's requirement.

Then in the linker script, put each array section in its own output
section with default (i.e. input section) alignment and assert that
it's not overaligned as a proxy for ensuring the output section
starts precisely at an object and there is no inter-object padding
beyond array-compatible alignment.

Fixed: 46972
Change-Id: I9eda6f3db2230158a9b6c5cfb37dd7f7685a62b0