[wlan][policy] change PSK length to 32
Fixed the length of PSK in policy layer: a PSK is 64 hexadecimal
characters long but when stored in bytes it should be stored as 32
bytes. Saving PSK is supported by new persistent storage (stash) but not
supported by legacy storage, so we will not write credentials saved as
PSK to legacy storage.

Network configs that were saved with a PSK before this change will
essentially be dropped. Stash has been modified to not fail loading if
it cannot understand something in stash - it will just ignore it.

Tests: Added stash test and modified unit tests:
	fx test wlancfg-tests
       Tested saving PSK through SL4F and manually tested
       auto connect behavior with PSK and persist storage.
Bug: 46995
Change-Id: If278b2fafc6c48c25b34bd89ddea764d0d99af2c