[dhcp] Apply client retransmission delay correctly
Retransmission delay is the delay until next retransmission if none of
the expected responses are received. E.g. from RFC2131 section 3.1,
for DHCPREQUEST "the client times out and retransmits the DHCPREQUEST
message if the client receives neither a DHCPACK or a DHCPNAK message."

Before this fix, *any* packet received on the wire restarts the
retransmission delay timer, causing the actual delay to be usually
longer than the delay configured.

Test:
fx run-test netstack_dhcp_gotests

Change-Id: I607f63b279b0a8ce320535199af0da7d34af5977