[wlan][policy] Match security types in policy
Interpret more protection/security types of scan results for connecting
to WEP networks. Previously we only distinguished between WPA2 and not,
so everything that wasn't WPA2 matched to None. Now we can match other
types of networks with the ones in the saved networks manager. For
example if we save a WEP network and see it in the scan results, we may
now try to connect to it. This should let us auto-connect to security
types beyond WPA2 and None. If a seen network is not compatible we will
not try to auto connect.

Bug: 46787
Tests: Added unit tests:
	fx run-test wlancfg-tests
       Manually tested that saving a WEP network through SL4F would
       cause an auto-connect.
Change-Id: Ied259c5a0f945a81b1c0426c991452023428f0e1