[gn sdk] Download specific instance id of Fuchsia DevTools
Previously we just used the latest CIPD binaries, which may not be
working. Also, the latest tag will not point to today's release at
some time in the future, so the checkout will not be repeatable.
We need to update the instance id manually as new stable releases
arrive, and this can be automated if DevTools has prebuilts in-tree.

Change-Id: I0ce0df1fa1189295bd547e487ea2315f26cb28e0