[fidl][c] Fix leak when rebinding async txn
Rebinding an async transaction using `fidl_async_txn_complete` is best
effort, and is implemented using an early exit on success. The contract
is for the function to always consume the async transaction. However we
were missing to free on the early exit path (i.e. when rebind succeeds).

To reproduce the issue

    fx set core.x64 \
        --with-base //garnet/packages/tests:zircon \
        --variant asan --args 'asan_default_options = "" zircon_extra_args = { asan_default_options = "" }'
    fx build
    fx qemu -kN

    > runtests -t fidl-simple-test

Test: fidl-simple-test
Bug: 46420
Change-Id: I19034beba67e1e1d3bc6b8282d01d46a45607ff2