[camera] Fix bug in DeviceTest
This change fixes a bug in the DeviceTest suite. Because the StreamImpl
is set to only allow one outstanding frame, and the test enqueues two
frames, the StreamImpl can (correctly) drop the first frame if the
client has not received it yet. This change fixes the test by only
sending the second frame after the client releases the first frame.

MULTIPLY: `[
  {
    "name": "camera_tests",
    "total_runs": 100
  }
]`

BUG: 47435

Test: repro'd issue locally, verified root cause with debugger
      verified no repro in 1000 runs with fix applied
Change-Id: I5a92879e82c1744f06687177b546be1b41452940