[lsan][trace] Suppress leak sanitizer on trace_integration_tests
Leak sanitizer detected bugs on the above tests. We suppress
the leak sanitizer on the leaky targets for now, and these leaks
need to be fixed to fully enable leak detector.

Note: the integration test invokes nested_environment_test and
basic_integration_test_app, so the LSan should be suppressed on
these binaries as well.

Fixed: 46468
Bug: 46850
Change-Id: I8120e53f478a8a63bdcb76d96fb4aaec16210d9b