[gn sdk] Always log femu.sh output and dump it out on error
Previously we did not save the log by default, and it is pretty easy
to get the args wrong and femu.sh will fail. Now we always log to a
temp file if necessary, and then dump it out on failure. If the run
succeeds then we delete the temp file and do not dump it.

Change-Id: I294bfd4a2b8a375a3c8b958dd567e7470a1a6471