[zircon][abr] Copy libabr temporarily to zircon
Move liabbr from src/firmware temporarily to zircon/system/ulib for use
in BootManager in the paver library. libabr/lib/BUILD.gn has been
modified to make it a zircon library. Unittests are also integrated into
the build system. The README file in the original location has been
revised to reflect that it is moved temporarily and will be restored
once we have unified build.

The CL is the predecessor of http://fxr/367464

Bug: 45121

Change-Id: I2db910698e13fa0abd919cae7b1488bd7ffa03b8