[scenic/escher] Revamp sampler creation/caching.
Step 7 in a series of CLs to build Vulkan pipelines at startup, rather
than lazily as needed.  This changes the way that Vulkan samplers are
obtained, so that the identical sampler will be available early on,
when we are warming the pipeline cache.

Fixed: 24574

Testing: added test for new SamplerCache class, which is also covered
by existing higher-level tests (e.g. pixel-tests).

Change-Id: I7648dd38de1123bdd5265f9672ec725823ac8261