[device-finder] error handling/debugging update
If non-fatal errors are encountered during poll, polling will be
attempted again. If a fatal error is encountered, `errno` will be
returned to the caller for debugging.

Invocation of the `%v` formatter for dns-sd.h API errors has been
removed and replaced with a `dnsSDError` type for easier debugging in
the event that interaction with the DNS-SD daemon fails.

Fixed: 47372

Change-Id: I813764191dce7839f2c0b452f663f8ee260b59a3