[kazoo] Move NewLineTrackingWriter's logic into the base Writer/StringWriter
This is a step towards combining Writer and StringWriter into a single
class, which I'll do in a follow-on change.

Rename MaybePrintln() to PrintSpacerLine() (hopefully a more
descriptive name).

Remove the CustomImplementation test.  It will go away when Writer and
StringWriter are combined, and I don't want to have to update it to
implement the temporarily-introduced virtual method.

Bug: none
Test: kazoo-test (unit test added)
Change-Id: I84a240668c6103da2d62f3c23b4fa6f69fd53934