[a11y] Adds interface class to allow a11y services request focus.
This change adds an interface only class (implementation in an upcoming cl),
that allows accessibility services to request a Focus Chain change with the
FocusChainManager.

The main advantage is that the ViewRef KoID can be used, which removes the need
to have a copy of the ViewRef to request focus.

This change adds the mock of this interface, which is used by services that
touch this interface in tests.

The change that migrates callers to use this interface will be sent only after
the implementation cl, so things are smaller and easier to review.

Change-Id: I67de19d79df35f891c42d2b7caac39f71f226863