[kstats] Add "-l" for graphical load bars.
Add a mode to kstats where we show graphical bars representing cpu load
for all processors on the system. The bars are 20chars wide and it shows
4 cpus per line. This should make a line width comparable to the wide of
"kstats -c". A little color formatting for easier viewing.

Ultimately something that redraws itself like a htop would be cool but
this gets us a little closer.

Looks like this rendered on qemu:
https://photos.app.goo.gl/UFezU1qY3JVEF1my5

Change-Id: Icf4c401474220be7a4afdbe131ea7f729179dfd5