[scenic][flatland] Move link properties to the UberStruct.
LinkProperties are still sent to the child Flatland instance as soon as the
link is resolved, but are otherwise treated as local to the Flatland instance
which created the child link. This means they are published to the UberStruct
like local topologies, and LinkSystem::UpdateLinks consumes the UberStructs
like all other global processing functions.

Fixed: 44334

TEST=fx run-test scenic_tests -t flatland_unittests

MULTIPLY: `[
  {
    "name": "scenic_tests",
    "os": "fuchsia",
    "total_runs": 40
  }
]`

Change-Id: I9957c582e845d061a07a94dedd2e967c8ba0cb00