[shell] Pass token text to terminal constructor
At present we just throw the text away, but with the new generic Token
parser, we will be able to define subclasses that don't throw it away.
The only requirement is that they all have to have the same constructor
signature, so even though the base Terminal class doesn't use it, it
still has to accept it.

Change-Id: Ic07913ff8d2f6b11aa05322fc4d0ef809e1d5fb4