[FDF][Multibind]Pass in component-device-id to protocols.
Multibindable components(shared resource) like power domains allow clients to
register/unregister for the power domains. When a composite device registers for
the power domain, it passes in the acceptable voltage range for this device. Generic power
driver that is managing the power domain, will add this device to the dependents
and notes down this voltage range. Thereafter, any voltage transitions
will be done so that the voltage is acceptable for all the devices that
are dependent on the power domain.

It is important to know which composite device is calling into
register/unregister. That will allow us to add/remove acceptable voltage
ranges and also register a client only once and not multiple times. Today,
the protocol ops, do not have knowledge of which composite device
is calling into them.

This change adds a mechanism to create a protocol context that is
private for each component device. The shared component
driver, will maintain child context for each of thes component devices
until the composite device is unbound. This change adds 2 protocol ops
to achieve this.
1. open_protocol_session_multibindable
2. close_prtocol_session_multibindable
Multibindable sharaed components that need a notion of which composite
device is the caller of protocol ops need to implement both of these
functions.

Bug: 36507

Test: runtests -t platform-bus-test

Change-Id: Iee9b78401b53701cc4c86a40d12ba75af429b1c3