[gpt] Support non-512 byte block sizes during GPT initialization.
We currently assume that blocksize is 512 bytes, and will fail GPT
initialization if this is not true.

Update the code to avoid attempting partial block writes, and to
avoid assuming that sizeof(mbr::Mbr) == blocksize.

This allows the following disabled tests to start passing:

  EfiDevicePartitionerTests.FlakyAddPartitionTooSmall
  EfiDevicePartitionerTests.FlakyInitializeWithoutFvmFails
  EfiDevicePartitionerTests.ValidatePayload
  SherlockPartitionerTests.FlakyAddPartitionNotSupported
  SherlockPartitionerTests.FlakyFindBootloader
  SherlockPartitionerTests.FlakyInitializeWithoutFvmFails

though they remain disabled due to fxb/42894.

Bug: 47409
Change-Id: Ida79f0957dbbfc378c0a578abc89556821c4df95