[blobfs] Factor out BlobLoader from Blob
Separates the responsibilities of loading blobs from disk (out of the
current monolith blob.cc).

The only functional change in this CL is that we now *always* map the
Merkle tree into a separate VMO from data. Previously we only did this
for the pager-backed case, but this led to complex logic.

Broadly there are three (current) cases for loading:
- Uncompressed, unpaged
  - Load blob from disk into client VMO
  - Verify contents
  - Give VMO for client
- Uncompressed, paged
  - Load blob from disk
  - Create paged VMO for client
  - Verify on-demand
- Compressed, unpaged
  - Load blob from disk into temporary VMO
  - Decompress into client VMO
  - Verify contents
  - Give VMO to client

Tested: New unit tests & existing blobfs test suite
Bug: 44742
Change-Id: I5535061833423a1ca69eaf70791f960dc2cbbca9