[a11y] Fix swipe gestures to handle move event.
This cl is allowing pointer events to be valid if the move event has the
same co-ordinates as down event.

Currently, while performing swipe gestures, there are back to back
pointer events where first event is "Down" at co-ordinate (x,y) and the
second event is "Move" at co-ordinate (x,y). This was causing the gesture
recognizer to fail in ValidateSwipeSlopeAndDirection().

Testing: * Build
  * Manually tested on device.
  * fx run-test a11y_lib_tests
Bug: 42835
Change-Id: Ia90dd0850041d265045573381456eb84f4859719