[fidl_codec] semantic for Launcher::CreateComponent
Defines HandleDescription for the semantic parser.

Adds a string value to ExpressionValue. Convert fidl_codec::Value to a
string when the fidl_codec value is a string.

Now the builtin semantic is used only if requested.

Splits the builtin semantic into several strings (one per library). This
way, we will be able to load the semantic only if we load the library.

Change-Id: I0edd4bb5ee3c235d69a97d9c8b97d10cd193c47f