[gn sdk] Adding test package rule template.
This template is used to create a gtest c/c++ executable that is
contained in a component and packaged for execution.

There is metadata added to the target so a generate_file() GN target
can be used to collect all the information about the tests into a
JSON file.

Follow-on CLs will handle how to process the metadata and run the tests.

Change-Id: I8a570c128902b4af969fd5c7964b6f1c26afc53b