[audio_core] Add Context class
The purpose of |Context| is to encapsulate all the global state that
previously lived inside of AudioCoreImpl. This also simplifies passing
around of many objects since we can just provide the context parameter
which can provide access to the RouteGraph, LinkMatrix, etc. Similarly
this allows us to make more effective unit tests by migrating from Fake,
Stub, or Mock objects to using the real versions by providing access to
a test Context instance in the test fixture.

This will simplify adding the fuchsia.ultrasound.Factory interface to
audio_core, which will be a peer to fuchsia.media.AudioCore but for
ultrasonic streams.

Test: fx run-test audio_core_unittests
      fx run-test audio_pipeline_tests
Change-Id: I0aabb8dd0d9c5f83059bf35ba7e7a24e1280029c