[benchmarking] Append buildbucket to master name when uploading to catapult
Currently multiple builders are sending the data to the same graph
because they share the same name. They only differ in the bucket name.

This CL adds the bucket name as part of the master.

By changing the master name, we break continuity for all results
uploaded by SL4F. We will attempt renaming the old results to match the
new names.

Bug:47178

TEST: fx set core.arm64 --with-base=//sdk/testing/sl4f/client:tests && fx build && fx run-host-tests sl4f_client_tests

Change-Id: I1dfe5d981c992338c2f9fdfc1e659c6ea1a0bab7