[audio_core] Add rate to audio_core_config.json
Update the config loader to support device profiles for both input and
output devices. For output profiles, we set a rate property on each mix
group. The root mix group will define the target rate that we'll attempt
to match with the hardware.

For input devices, we add a single 'rate' property that will be the
target rate for the input device. These properties are not yet wired up
with the drivers to actually request these rates.

Bug: 42541
Test: fx run-test audio_core_unittests
Change-Id: I051722ce1a531a41e067a0e70d67dc216fa8c960