[gn sdk] Download specific aemu prebuilt and support multiple versions
Previously we just used the latest CIPD binaries, which may not be working.
Now, we use aemu.version, which contains the latest working version. If this
file is changed, then we automatically download and use that aemu prebuilt.
The aemu.version file is currently manually updated, but we will eventually
auto-update this as new GN SDKs are being created.

Change-Id: I9644b90e11e1063bedfd4735a800cc7383d3efc6