[a11y] Adds an interface class that allows consuming semantics.
This change adds an interface only class for accessibility services query
existing semantic data. The implementation (inside ViewManager), will be sent in
an upcoming cl.

With this interface:
1. Screen Reader will be able to have access only to the methods to consume
semantics, and will not need to access ViewManager's full set of capabilities.
2. Focus Chain manager, which needs to know which Views are providing semantics,
will be able to query this information.

This change adds a mock for this interface, that will be used by (1 and 2),
listed above.

The change to migrate callers to use this interface will be done in a separate cl.

Change-Id: I0265910b78c1ce05f718842d31a9b2335abe2543