[installer] Add support for paving kernel images
This CL adds support for the installer to read
partitions into memory, and then send them to the paver
to be written to disk. Currently, only kernel images are paved.
The A image is paved to both the A and B slots, and the recovery
image is paved to the recovery slot.

Bug: 44595
Change-Id: Ia0960e3b9eb7effd9331971a4e3d8bd98657edc8