[omaha client] Cap the duration between update checks at 60min
In the case that the current time and the last_update_check are such
that one hour after the last_update_check is more than one hour into the
(supposed) future, cap the duration at 1 hour.

Extend the MockTimer functionality to assert that all expected waits
have been waited on.

Bug: 46833
Test: Manual reboot testing and log analysis
Test: fx run-test omaha-client-lib-tests

Change-Id: I14794b6a47f638ab13595107caee87e4ec4d2ba3