[inspect] Adds a new health metric
The new health metric is the monotonic timestamp of the time when the
health node has been instantiated.   This provides an easy way to know
when the binary has been started, and can offer additional insight into
the component lifecycle.

Changes atomically the C++ and rust implementations.  dart
implementation is committed separately, but should be equivalent.

Modifies higher-level tests not to depend on the specific structure of
Health nodes that they do not necessarily care about.

Bug: 47295
Change-Id: I947216e70c237b25feb662718b16db2fc9ecc2c3