[media] clock utilities
This CL adds two sets of clock-related utilities for use by
audio services, utilities and drivers. The first makes it
easier to work with zx::clock objects and to convert between
affine transforms (used by zx::clock) and TimelineFunction
(used across media). The second is a simple PID controller,
which will be used to smooth the adjustment of clock rates.

Test: audio_clock_utils_unittests contains 19 new unittests

Change-Id: Iffbdf0629aa85992c74ab359ccb37bb86ccff1c5