[ffx] mDNS pipeline hook-up.
This adds in the mDNS client pipeline to ffx. This includes some basic
configuration structs and data structures for keeping track of found
devices. This is mainly focused on dropping in the mDNS pipeline rather
than handling which data structures to send to the user.

Current implementation sends an mDNS broadcast every 120 seconds to
check for devices.

Fixes: 47737

Change-Id: I9b1d0307e943e4d10f6fbb628188d64a46af7ca9