[a11y] Implements Focus Chain Requests in Focus Chain Manager.
This change implements the possibility of Focus Chain manager request focus
chain updates. It will query the ViewRefs that are providing semantics via the
SemanticsSource interface and request with Scenic a Focus Chain update if the
View is providing semantics.

A11yFocusManager will ask for Focus Chain Updates via
AccessibilityFocusChainRequester interface, and this will be implemented in an
upcoming cl.

Change-Id: I8d6aa3613d069c3fdb99e9700578c9951c84b966