[brcmfmac] Add the ability to create periodic timers.
If a timer is set as a periodic timer, the timer handling routine
calls the equivalent of timer_set() again with the specified delay. In
addition, the changes ensure that calling timer_set() or timer_stop()
from the timer handler does not cause any issues.

Test: Added SIM tests to verify the new feature works as well
as ensure there are no side-effects to other timers.

Change-Id: If15e0d32148a8341e0bcad438afc09e8b987628a