[runtests] Drop support for referring to components via pkgfs
Before this change, runtests would accept /pkgfs/packages/ paths
and translate them into component URIs. Now it requires the URIs
be supplied directly.

This is a request from the software delivery team.
IIUC, the intent is to remove all references to /pkgfs/packages/.

In the process, move the check for the presence of the component
executor out of SetUpForTestComponent() to make SetUpForTestComponent()
more easily testable. Newly added test uncovered a bug, which is also
fixed in this commit:
fbl::count_of() counts the null terminator, and thus the call

`strncmp(test_path, kPkgPrefix, fbl::count_of(kPkgPrefix))`

was never returning 0.

Bug: 43448
Change-Id: Ifb9ed0a31a2aa0be13119ed4d921638ead5efcba