[add-update-source] enable custom addresses
The --type flag did not work before, as --type localhost set repo_url
in the code, but that parameter was not used later on, so it was
ineffective.

--addr enables a user to specify the address the target will use to
connect to the package server and when used this skips all checks for
local-ness - this can be used with non-local servers for example when
port-forwarded over ssh.

Test: Device previously nearby, but since offline:
% fx add-update-source
ssh: connect to host fe80::...%en0 port 22: Operation timed out
ERROR: unable to determine host address as seen from the target.  Is the target up?

Test: successful addition:
% fx add-update-source
% fx shell pkgctl repo
fuchsia-pkg://devhost

Test: fx serve is not running:
% fx add-update-source
2020/03/06 18:42:12 error adding source: failed to GET file: Get "http://[fe80::...%25ethp0014]:8083/config.json": dial tcp [fe80::...%ethp0014]:8083: errno(111)
ERROR: Unable to register update source.
ERROR:  - Is 'fx serve' or 'fx serve-updates' running?
ERROR:  - Can the target reach http://[fe80::...%25ethp0014]:8083/config.json ?

Test: A forwarded port:
% fx shell -R 8080:localhost:8083
$
% fx add-update-source --addr 127.0.0.1 --port 8080
% fx shell pkgctl repo
fuchsia-pkg://127_0_0_1
fuchsia-pkg://devhost
% fx shell pkgctl repo -v | grep mirror_url
        "mirror_url": "http://127.0.0.1:8080",
        "mirror_url": "http://[fe80::...%25ethp0014]:8083",

Change-Id: I30daa7ae88c647927de8af87ee4d672dfcf3e947