[vulkan] Fix imagepipe swapchain's display controller probing.
This small CL fixes how the display controller device is probed
by the Vulkan full-screen swapchain implementation.

Instead of assuming that /dev/class/display-controller/000 is
always the correct path for the device file, probe the directory
for existing entries and use the first one found there.

This fix removes an annoying flake that happens when the
display controller dies for some reason, because it will then
be restarted with a new name (e.g. 001), making the previous
code fail unexpectedly.

+ Add a new unit-test to verify Vulkan presentation surface
  creation, with or without the framebuffer.

Bug: 45919
Testing: fx run-test image-pipe-swapchain-tests

Change-Id: Ia7d54beaa2473d3522cdc3bdc945434c0e6eb982