[fuzzing][rust] Implement #[fuzz] proc_macro.
This CL replaces the `declare_fuzzers!` macro with a `#[fuzz]`
proc_macro. The new proc_macro takes a fuzz target function, and
cooperates with the `rustc_fuzzer` GN template to conditionally
compile a single fuzzer at a time. It also allows for more
flexible parameters: in addition to a `&[u8]` slice, it allows
fuzz target functions to one or more parameters of any type with
the `Arbitrary` trait.

Bug: 24965

Change-Id: If3521ab73961fe3346b004eb73e4983f48d95390