[netstack3] Configure UDP ICMP error responses at runtime
Previously, we used a conditional compilation flag in order to
configure whether or not the stack sent ICMP Port Unreachable
messages in response to inbound UDP packets. This made it difficult
to test, and meant that some of our testing code (which was also
behind a conditional compilation flag) had bitrotted.

This change makes this configuration a runtime option, and allows
us to always test the behavior by having the relevant test simply
configure the stack to send such replies, though the default is
still to not send them.

Test: Update `ip::icmp::tests::test_port_unreachable` and make it
      compile unconditionally; run with:

      fx run-test netstack3_tests -t netstack3_core_lib_test

Change-Id: Iacf4657e9f1c65275195c9af8a68f88763aa69aa