[session] Clarify coordinates across systems, part 5
Input and scenic are just two examples of systems that want to use
different corrdinate units within Fuchsia. This CL creates a struct that
allows systems to set and retrieve coordinates in pixels, pips, or
millimeters.

This CL removes unused code that was temporarily added for the
transition and fixes the display density for Pixelbook Go.

Bug: 44272
Bug: 47385
Change-Id: I54effe20725389a4eb65f8f2602028e9dcbb1063