[kernel][x86] Add interfaces to disable Turbo/Core Performance Boost
Turbo Boost (Intel) and Core Performance Boost (AMD) are mechanisms that
allow a processor to exceed nominal frequency or utilize 'free' TDP or
design current to offer more performance in the short term. On Intel
client SoCs, these mechanisms operate in the millisecond interval,
which makes them very effective for bursty/interactive applications.

Certain applications may wish to disable turbo for performance
consistency; introduce an in-kernel interface to do so and provide an
implementation for Intel processors.

One intended use for this mechanism is the perfcompare bots, which
would be able to provide narrower CIs for microbenchmarks when Turbo
is disabled.

Change-Id: I64f2f002a4a039bebb0271db5e4b21a512485a0a