[test_util] Improve asserts to debug flaky test
I made a dumb error in my asynchronous test that only failed under QEMU.
The Counter Mutex is only locked for the duration of each function, so
inc() and get() were under separate locks, allowing another thread to
interrupt and call inc() between the first thread's calls to inc() and
get(). The resulted in sequences like: 3, 3, 5, 5, 6, 7, 7, 8, 10, 10.
The final count was always right.

I modified inc() for the benefit of anyone wanting to increment the
counter and receive each new value (from the same Mutex lock()). Using
this new result, the test now works as expected.

Comments and documentation were also improved to highlight the potential
effects on Counter instances in asynchronous situations, and the
guarantees.

Fixed: 47786

Test: fx run-test rust_test_util_tests -r 100 -- --nocapture async_count
Change-Id: I5bdf8b1d19874e29558e31c04c57583036602cd9