[banjo] Add duplicate libraries temporarily
Component driver needs dependencies on these libraries to override
DdkGetProtocol. In recent unification changes, these libraries moved to
GN build. Ideally the component driver needs to move to GN as well, but
the cannot be readily done as zbi_test that is in ZN build depends on
component driver. As a temporary solution, this change creates duplicate
zircon libraries that refer to the banjo file from sdk/banjo/*.

Change-Id: I71807be2fa9d0f7123514ba38c20c30972a752e8