[board][x86][cros] Avoid using invalid report IDs in HID descriptors.
We currently generate invalid HID descriptors for the Chrome OS EC
motion sensors: in particular, we generate descriptors with report ID 0,
which is reserved by the HID/USB spec.

Update the reports, and add a test case that ensures that generated HID
descriptors can be successfully parsed by the hid-parser library.

Bug: 47988
Change-Id: I250dd804fd16da80b8289c16e6fa6e7526b051bf