[netstack] suppress a netstack warning.
We have been seeing many warnings filling up syslog lately because
OnInterfacesChanged() fails with ErrPeerClosed or ErrBadHandle.
ErrPeerClosed would mean that the peer closed its side of the channel.
ErrBadHandle suggests a bug in the Go bindings - we're attempting to
send an event on a channel we ourselves have already closed.

Bug: 47328
Change-Id: I84736803d8648326fc8e2116289579b9fec65013