[system-tests] Use the base-package "update" binary
We don't have merkle pinning working yet for the base packages, so
running "update" ends up resolving the update package from the
upgrade repository, which might be incompatible with the current system.
To protect against this, this patch changes the TriggerSystemOTA() method
to determine the base package merkle for "update-bin/0", and explicitly
run /pkgfs/versions/$merkle/bin/update to make sure we are using the
correct version.

Test: Manually confirmed the upgrade test works.

Change-Id: Id8166d389dfccb0aa83a70054d598fd02bfb848b