[components] Install hooks on BuiltinEnvironment construction
Hook installation was scattered between BuiltinEnvironment construction
and setting up ServiceFs. Hooks installation had no business being in
ServiceFs setup, and so this CL moves all hooks installation to
construction instead. This CL has the added benefit of exercising the Hub
and the Events system more thoroughly in all our unit tests.

This is a prerequisite to using the Discover event on the root.

Change-Id: Id641701d786e6007f4bf200232b4668c0d6409b6