[wlan][policy] Pass in SME to policy client
This CL temporarily passes in a reference to the SME client into the
new Policy API handler. This temporary workaround will be obsoleted
by the Interface Manager that sakuma@ is working on.

For the moment, this workaround is necessary so that the new Policy API
can be exercised in automated testing (SL4F)

Test: - set_sme() is used in several tests in fxr/364139
      - act.py -c <path-to-config> -tc PolicyScanTest
        (depends on: https://r.android.com/1247112)
Change-Id: I1af447a9646cd6aae5e5a2556f13e1654b0380d9