[paver] add content type to DevicePartitioner
Adds a content type parameter to DevicePartition methods so that
different content can be sent to the same partition and handled
appropriately.

This will be used for cases like:
 * Sending multiple bootloader pieces (e.g. BL2 and TPL) to the
   same logical kBootloader partition
 * Switching to a new image format or new backend paving implementation

This just adds the capabilities to handle content types in the
DevicePartition API, and doesn't have any effect yet. A follow-on CL
will enable this by passing the content type through the paver FIDL.

Bug: 45606
Change-Id: Icd44e8917c76b1f61973c4826f5e29b055117b08