[wlancfg] Allow log::info! to show up in klog
Currently, logging is only enabled with sys_log. As a result,
messages printed with `info!` are missing in klog.

When there is no connectivity, the only logs available are serial logs,
which only include klog messages.

To help with debugging, use set_logger to allow logs to show up in both
sys_log and klog.

Bug: 47966
Test: Logging only change.
As a sanity check, verified that logs are indeed shown in klog.

Change-Id: I5a19d25a26933f199e27e81379e41164ad35a563