[feedback] Renable tests - previously flaky
We had tests flaking on x64 bots because the channel connecting the test
environemnt to mock cobalt was closing and fake cobalt wasn't resetting
mock cobalt at the end of tests. We now check-fail that we reset mock
cobalt and produce a more informative error message in case the flake
ever happens again.

TESTED=This CL has been run on trybots and has not exhibited flakes.

MULTIPLY: `[
  {
    "name": "feedback_agent_integration_test",
    "total_runs": 100
  },
  {
    "name": "crashpad_agent_integration_test",
    "total_runs": 100
  }
]`

Fixed: 47157

Change-Id: Ifce160b9a2803c2ec3065e237393c9cbbd6f8801