[zircon][unittest] fix DEATH_TEST memory leak
DEATH_TEST was leaking memory since the std::thread created to run the
death function would crash (intentionally) and never return to clean up
its resources.

This CL switches to use C thread APIs instead. This fixes the memory
leak so that we now pass with LSAN, but it's not a complete fix - the
dead thread still never cleans up it's resources e.g. VMO/VMAR mapping
for the thread stack.

This is OK because it's test-only so things will get cleaned up when the
process exits, and it would add more complexity than it's worth to try
to properly clean everything up.

Bug: 47448
Test: LSAN build no longer complains about leaks
Change-Id: I986a9fd9335cb182c6c3b395fcbc7e6ac9769fe4