[feedback] Introduce DeviceIdProvider
The Feedback device id utilities (InitializeDeviceId()) are now owned
by the DeviceIdProvider and so the DeviceIdProvider is the entity that
intializes the Feedback device id. This doesn't fix the fact that the
device id needs to be initialized before the data store is created, but
it makes it an implementation detail of feedback agent as opposed to it
happening in main().

TESTED=`fx run-test feedback_agent_tests`
TESTED=Ran `fx bugreport` with and without the change and the device id
didn't change.

Bug: 42590

Change-Id: I69cddf046f91adc4983b64b79f64c9772a711ec6