[identity] Rename test package with dashes
Recently the build system began enforcing
a rule to disallow underscores in package
names.  While the remainder of the identity
packages were added to a list of exceptions,
this package was missed (presumably because
it isn't built unless explicitly added).
This updates the package name to use dashes
to enable building and running the test.

Longer term, we should have a conversation
around what should have dashes vs underscores
in tree.  For now, this makes minimal changes
to enable building the test package.

Testing: fx test storage-manager-integration-test-pkg

Change-Id: Ia91c8a54285c914040a5a41790daa6c3e1f0c880